% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.sm-package}
\alias{mlr3learners.sm}
\alias{mlr3learners.sm-package}
\title{mlr3learners.sm: Learners from the {sm} package for 'mlr3'}
\description{
Adds learners from the {sm} package to 'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
